pyproject.toml
[project]
name = "recipe-items"
version = "0.1.0"
description = "Use AI to extract recipe ingredients from recipe websites"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "nest-asyncio>=1.6.0",
    "openai>=1.65.0",
    "pydantic>=2.10.6",
    "pydantic-ai>=0.0.29",
    "python-dotenv>=1.0.1",
    "typer>=0.15.2",
]
